Improve docs
fix links and move stuff around so rust doc looks ok.
